fuck nixos-generators
It turns out that using nixos-generators in our case is a bit stupid
because nixos-generators injects extra modules into the build, and those
modules are not visible to deploy-rs

This commit removes the dependence on nixos-generators by specifying all
relevant options (that would otherwise be injected by nixos-generators)
locally. It turns out that for non-qemu hosts this change is quite
minimal, but it fixes weird deploy-rs failures.